requirements.txt
streamlit
pandas
google-generativeai
altair
pydantic
pydantic-settings
pandera
pytest
python-dotenv
